requirements_hw_sensors.txt
# Items here are only checked after attempting to install the local version in the DEB itself
# Sensor HW Requirements
spidev
RPi.GPIO
gpiozero
pyserial
pyusb
# Other requirements related to the sensors
Pillow
pycparser
pyftdi
colorzero
zope.event
zope.interface
Brotli
cffi
retrying
